Refactor test utils
Remove setup_superuser and use the UserFactory directly.

Move get_parameter_values to utils instead of implementing it twice.

Fix typo in test.

Refs. TS-2597
